refactor(video): hoist ShaderEffect + effects to a backend-neutral home
Checkpoint 1 of the dual-language effect arc: ShaderEffect and the 19
built-in effect classes move from src/video/webgl/ to src/video/effects/,
and the pure GLSL source assembly (builtin parsing + vertex/fragment
boilerplate) is extracted verbatim into glsl_realization.js so the class
can dispatch per backend in the next step.

Zero behavior change, proven two ways: every existing shader/effect spec
passes unmodified, and a new generated-GLSL golden spec pins the
assembled sources byte-identical across the move (snapshots generated
against the pre-refactor assembly). Only mechanical path updates outside
the moved files (index.ts, loader parser, drawLight import, camera2d,
one dynamic-import path in lights.spec).
