+++ packages/melonjs/src/index.ts
@@ -62,6 +62,25 @@ import save from "./system/save.ts";
 import timer from "./system/timer.ts";
 import Tween from "./tweens/tween.ts";
 import CanvasRenderer from "./video/canvas/canvas_renderer.js";
+import BlurEffect from "./video/effects/blur.js";
+import ChromaticAberrationEffect from "./video/effects/chromaticAberration.js";
+import ColorMatrixEffect from "./video/effects/colorMatrix.js";
+import DesaturateEffect from "./video/effects/desaturate.js";
+import DissolveEffect from "./video/effects/dissolve.js";
+import DropShadowEffect from "./video/effects/dropShadow.js";
+import FlashEffect from "./video/effects/flash.js";
+import GlowEffect from "./video/effects/glow.js";
+import HologramEffect from "./video/effects/hologram.js";
+import InvertEffect from "./video/effects/invert.js";
+import OutlineEffect from "./video/effects/outline.js";
+import PixelateEffect from "./video/effects/pixelate.js";
+import ScanlineEffect from "./video/effects/scanline.js";
+import SepiaEffect from "./video/effects/sepia.js";
+import ShaderEffect from "./video/effects/shadereffect.js";
+import ShineEffect from "./video/effects/shine.js";
+import TintPulseEffect from "./video/effects/tintPulse.js";
+import VignetteEffect from "./video/effects/vignette.js";
+import WaveEffect from "./video/effects/wave.js";
 import { Batcher } from "./video/gpu/batcher.js";
 import {
 	isPortableTopology,
@@ -83,26 +102,7 @@ import Texture2d from "./video/texture/texture2d.ts";
 import { WebGLBatcher } from "./video/webgl/batchers/batcher.js";
 import PrimitiveBatcher from "./video/webgl/batchers/primitive_batcher.js";
 import QuadBatcher from "./video/webgl/batchers/quad_batcher.js";
-import BlurEffect from "./video/webgl/effects/blur.js";
-import ChromaticAberrationEffect from "./video/webgl/effects/chromaticAberration.js";
-import ColorMatrixEffect from "./video/webgl/effects/colorMatrix.js";
-import DesaturateEffect from "./video/webgl/effects/desaturate.js";
-import DissolveEffect from "./video/webgl/effects/dissolve.js";
-import DropShadowEffect from "./video/webgl/effects/dropShadow.js";
-import FlashEffect from "./video/webgl/effects/flash.js";
-import GlowEffect from "./video/webgl/effects/glow.js";
-import HologramEffect from "./video/webgl/effects/hologram.js";
-import InvertEffect from "./video/webgl/effects/invert.js";
-import OutlineEffect from "./video/webgl/effects/outline.js";
-import PixelateEffect from "./video/webgl/effects/pixelate.js";
-import ScanlineEffect from "./video/webgl/effects/scanline.js";
-import SepiaEffect from "./video/webgl/effects/sepia.js";
-import ShineEffect from "./video/webgl/effects/shine.js";
-import TintPulseEffect from "./video/webgl/effects/tintPulse.js";
-import VignetteEffect from "./video/webgl/effects/vignette.js";
-import WaveEffect from "./video/webgl/effects/wave.js";
 import GLShader from "./video/webgl/glshader.js";
-import ShaderEffect from "./video/webgl/shadereffect.js";
 import WebGLRenderer from "./video/webgl/webgl_renderer.js";
 import WebGPUPrimitiveBatcher from "./video/webgpu/batchers/primitive_batcher.js";
 import WebGPUQuadBatcher from "./video/webgpu/batchers/quad_batcher.js";
