+++ packages/melonjs/src/video/effects/chromaticAberration.js
@@ -1,4 +1,4 @@
-import ShaderEffect from "../shadereffect.js";
+import ShaderEffect from "./shadereffect.js";
 
 /**
  * A shader effect that offsets the RGB color channels to create a
