+++ packages/melonjs/src/video/effects/dissolve.js
@@ -1,4 +1,4 @@
-import ShaderEffect from "../shadereffect.js";
+import ShaderEffect from "./shadereffect.js";
 
 /**
  * A shader effect that dissolves the sprite using a noise-based threshold.
