+++ packages/melonjs/src/video/effects/flash.js
@@ -1,4 +1,4 @@
-import ShaderEffect from "../shadereffect.js";
+import ShaderEffect from "./shadereffect.js";
 
 /**
  * A shader effect that flashes the sprite with a solid color.
