+++ packages/melonjs/src/video/effects/invert.js
