+++ packages/melonjs/src/video/effects/outline.js
@@ -1,4 +1,4 @@
-import ShaderEffect from "../shadereffect.js";
+import ShaderEffect from "./shadereffect.js";
 
 /**
  * A shader effect that draws a colored outline around the sprite.
