+++ packages/melonjs/src/video/effects/scanline.js
@@ -1,4 +1,4 @@
-import ShaderEffect from "../shadereffect.js";
+import ShaderEffect from "./shadereffect.js";
 
 /**
  * A shader effect that overlays horizontal scanlines on the sprite.
