+++ packages/melonjs/src/video/effects/sepia.js
