+++ packages/melonjs/src/video/effects/shadereffect.js
@@ -5,114 +5,8 @@ import {
 	on,
 } from "../../system/event.ts";
 import Texture2d from "../texture/texture2d.ts";
-import GLShader from "./glshader.js";
-import quadVertex from "./shaders/quad.vert";
-
-/*
- * ---- Shader builtins -------------------------------------------------------
- *
- * Inside a ShaderEffect fragment body, three names get special treatment so
- * users never compute screen/frame UVs by hand:
- *
- * - `uniform sampler2D <name> : screen_texture;` — the engine strips the
- *   annotation and keeps the sampler filled with a capture of everything
- *   drawn so far (a back-buffer copy, via the renderer's shared
- *   toFrameTexture slot). An optional wrap mode is accepted:
- *   `: screen_texture(repeat)`.
- * - `screen_uv`  — varying with this fragment's position in that capture
- *   (0..1 across the screen).
- * - `noise_uv`   — varying with a frame-local coordinate across the drawn
- *   object (undoes atlas packing; scaled to object pixels so patterns keep
- *   their density when the destination is scaled).
- *
- * The builtins only activate when referenced, and never when the body
- * carries its OWN declaration of the identifier — a shader that used these
- * names before this feature keeps compiling unchanged.
- */
-const SCREEN_TEXTURE_ANNOTATION =
-	/\buniform\s+sampler2D\s+([A-Za-z_]\w*)\s*:\s*screen_texture(?:\((repeat|repeat-x|repeat-y|no-repeat)\))?\s*;/g;
-const SCREEN_UV_IDENTIFIER = /\bscreen_uv\b/;
-const NOISE_UV_IDENTIFIER = /\bnoise_uv\b/;
-
-/**
- * whether the body declares `name` itself (as a varying/uniform/attribute) —
- * the engine then leaves that identifier fully user-managed
- * @ignore
- */
-function hasOwnDeclaration(source, name) {
-	return new RegExp(
-		`\\b(?:varying|uniform|attribute)\\s+\\w+\\s+${name}\\s*;`,
-	).test(source);
-}
-
-/**
- * Parse the builtin usages out of a fragment body: collect and strip the
- * `: screen_texture` annotations, and detect the `screen_uv` / `noise_uv`
- * varyings. Bodies without builtins pass through byte-identical.
- * @ignore
- */
-function parseShaderBuiltins(fragmentBody) {
-	const screenTextures = [];
-	const body = fragmentBody.replace(
-		SCREEN_TEXTURE_ANNOTATION,
-		(match, name, repeat = "no-repeat") => {
-			screenTextures.push({ name, repeat });
-			// keep the plain `uniform sampler2D <name>;` declaration
-			return match.replace(/\s*:\s*screen_texture(?:\([\w-]+\))?/, "");
-		},
-	);
-	const screenUV =
-		(screenTextures.length > 0 || SCREEN_UV_IDENTIFIER.test(body)) &&
-		!hasOwnDeclaration(body, "screen_uv");
-	const noiseUV =
-		NOISE_UV_IDENTIFIER.test(body) && !hasOwnDeclaration(body, "noise_uv");
-	return { body, screenTextures, screenUV, noiseUV };
-}
-
-/**
- * The effect's vertex shader: the stock quad template, or — when a builtin
- * varying is in play — a variant that additionally computes `screen_uv`
- * (clip space → 0..1) and/or `noise_uv` (atlas UV → frame-local, fed by
- * `_setNoiseUVRect` through the `ME_*` uniforms). Built deterministically
- * from a template; user source is never rewritten.
- * @ignore
- */
-function buildEffectVertex(builtins) {
-	if (!builtins.screenUV && !builtins.noiseUV) {
-		return quadVertex;
-	}
-	return [
-		"attribute vec3 aVertex;",
-		"attribute vec2 aRegion;",
-		"attribute vec4 aColor;",
-		"uniform mat4 uProjectionMatrix;",
-		"varying vec2 vRegion;",
-		"varying vec4 vColor;",
-		...(builtins.screenUV ? ["varying vec2 screen_uv;"] : []),
-		...(builtins.noiseUV
-			? [
-					"uniform vec2 ME_size_obj;",
-					"uniform vec2 ME_size_img;",
-					"uniform vec2 ME_offset;",
-					"varying vec2 noise_uv;",
-				]
-			: []),
-		"void main(void) {",
-		"    vec4 ME_clip = uProjectionMatrix * vec4(aVertex, 1.0);",
-		"    gl_Position = ME_clip;",
-		...(builtins.screenUV
-			? ["    screen_uv = ME_clip.xy / ME_clip.w * 0.5 + 0.5;"]
-			: []),
-		...(builtins.noiseUV
-			? [
-					"    noise_uv = aRegion * (ME_size_img / ME_size_obj) - ME_offset / ME_size_obj;",
-				]
-			: []),
-		"    vColor = vec4(aColor.bgr * aColor.a, aColor.a);",
-		"    vRegion = aRegion;",
-		"}",
-	].join("\n");
-}
+import GLShader from "../webgl/glshader.js";
+import { buildGLSLProgram } from "./glsl_realization.js";
 
 /**
  * A simplified shader class for applying custom fragment effects to renderables.
@@ -230,10 +124,10 @@ export default class ShaderEffect {
 			return;
 		}
 
-		// Shader builtins: parse & strip `: screen_texture` annotations,
-		// detect the free `screen_uv` / `noise_uv` varyings (see the module
-		// header). A body using none of them passes through untouched.
-		const builtins = parseShaderBuiltins(fragmentBody);
+		// assemble the GLSL sources (builtin parsing + boilerplate) — the
+		// pure-assembly half lives in glsl_realization.js, pinned
+		// byte-identical by the generated-GLSL golden spec
+		const program = buildGLSLProgram(fragmentBody);
 
 		/**
 		 * samplers annotated `: screen_texture` — the renderer checks this to
@@ -241,29 +135,15 @@ export default class ShaderEffect {
 		 * @type {Array<{name: string, repeat: string}>}
 		 * @ignore
 		 */
-		this._screenTextureUniforms = builtins.screenTextures;
+		this._screenTextureUniforms = program.screenTextures;
 		/** @ignore */
-		this._hasNoiseUV = builtins.noiseUV;
-
-		// wrap the user's apply() with the texture-sampling boilerplate
-		const fragment = [
-			"uniform sampler2D uSampler;",
-			"varying vec4 vColor;",
-			"varying vec2 vRegion;",
-			...(builtins.screenUV ? ["varying vec2 screen_uv;"] : []),
-			...(builtins.noiseUV ? ["varying vec2 noise_uv;"] : []),
-			builtins.body,
-			"void main(void) {",
-			"    vec4 texColor = texture2D(uSampler, vRegion) * vColor;",
-			"    gl_FragColor = apply(texColor, vRegion);",
-			"}",
-		].join("\n");
+		this._hasNoiseUV = program.noiseUV;
 
 		/** @ignore */
 		this._shader = new GLShader(
 			renderer.gl,
-			buildEffectVertex(builtins),
-			fragment,
+			program.vertex,
+			program.fragment,
 			precision || renderer.shaderPrecision,
 		);
 		this.enabled = true;
@@ -279,7 +159,7 @@ export default class ShaderEffect {
 		// wire every annotated screen sampler to the renderer's shared frame
 		// capture — a live GPU-resident entry, re-bound fresh on each draw and
 		// skipped while no capture has been taken yet
-		for (const screenTexture of builtins.screenTextures) {
+		for (const screenTexture of program.screenTextures) {
 			this.setTexture(
 				screenTexture.name,
 				renderer.getSharedFrameTexture(),
