+++ packages/melonjs/src/video/effects/shine.js
@@ -1,4 +1,4 @@
-import ShaderEffect from "../shadereffect.js";
+import ShaderEffect from "./shadereffect.js";
 
 /**
  * A shader effect that sweeps a bright highlight band across the sprite —
