+++ packages/melonjs/src/video/effects/tintPulse.js
@@ -1,4 +1,4 @@
-import ShaderEffect from "../shadereffect.js";
+import ShaderEffect from "./shadereffect.js";
 
 /**
  * A shader effect that pulses a color overlay on the sprite.
