+++ packages/melonjs/src/video/effects/vignette.js
@@ -1,4 +1,4 @@
-import ShaderEffect from "../shadereffect.js";
+import ShaderEffect from "./shadereffect.js";
 
 /**
  * A shader effect that darkens the edges of the screen, drawing focus
