+++ packages/melonjs/src/video/effects/wave.js
@@ -1,4 +1,4 @@
-import ShaderEffect from "../shadereffect.js";
+import ShaderEffect from "./shadereffect.js";
 
 /**
  * A shader effect that applies a sine wave distortion to the sprite.
